+++ lib/settlement/driip-settlement-challenge-contract.js
@@ -1,10 +1,119 @@
 'use strict';
 
+const {isRevertContractException} = require('./utils');
 const NahmiiContract = require('../contract');
+const NestedError = require('../nested-error');
 
 class DriipSettlementChallengeContract extends NahmiiContract {
     constructor(walletOrProvider) {
         super('DriipSettlementChallengeByPayment', walletOrProvider);
+    }
+
+    /**
+     * Returns the proposal nonce
+     * @param {Address} address - The wallet address
+     * @param {Address} ct - The currency address
+     * @param {Integer} id - The currency id
+     * @returns {Promise} A promise that resolves into a BigNumber value representing the nonce of the latest challenge.
+     * @example
+     * let nonce = await driipSettlement.getCurrentProposalNonce(address, ct, id);
+     */
+    async getCurrentProposalNonce(address, ct, id) {
+        try {
+            return await this.proposalNonce(address, ct, id);
+        }
+        catch (err) {
+            if (isRevertContractException(err)) 
+                return null;
+            
+            throw new NestedError(err, 'Unable to get the nonce of the current proposal.');
+        }
+    }
+
+    /**
+     * Returns the expiration time of the proposal
+     * @param {Address} address - The wallet address
+     * @param {Address} ct - The currency address
+     * @param {Integer} id - The currency id
+     * @returns {Promise} A promise that resolves into a BigNumber value representing the timeout timestamp.
+     * @example
+     * let expirationTime = await driipSettlement.getCurrentProposalExpirationTime(address, ct, id);
+     */
+    async getCurrentProposalExpirationTime(address, ct, id) {
+        try {
+            return await this.proposalExpirationTime(address, ct, id);
+        }
+        catch (err) {
+            if (isRevertContractException(err)) 
+                return null;
+            
+            throw new NestedError(err, 'Unable to get the expiration time of the current proposal.');
+        }
+    }
+
+    /**
+     * Returns expire state of the current proposal
+     * @param {Address} address - The wallet address
+     * @param {Address} ct - The currency address
+     * @param {Integer} id - The currency id
+     * @returns {Promise} A promise that resolves into a boolean value indicating if the latest challenge has expired.
+     * @example
+     * let hasExpired = await driipSettlement.hasProposalExpired(address, ct, id);
+     */
+    async hasProposalExpired(address, ct, id) {
+        try {
+            return await this.hasProposalExpired(address, ct, id);
+        }
+        catch (err) {
+            if (isRevertContractException(err)) 
+                return null;
+            
+            throw new NestedError(err, 'Unable to check whether the proposal has expired.');
+        }
+    }
+
+    /**
+     * Returns intended stage amount of the current challenge proposal
+     * @param {Address} address - The wallet address
+     * @param {Address} ct - The currency address
+     * @param {Integer} id - The currency id
+     * @returns {Promise} A promise that resolves into a BigNumber value representing the indended stage amount for the latest challenge.
+     * @example
+     * let stagedAmount = await driipSettlement.getCurrentProposalStageAmount(address, ct, id);
+     */
+    async getCurrentProposalStageAmount(address, ct, id) {
+        try {
+            return await this.proposalStageAmount(address, ct, id);
+        }
+        catch (err) {
+            if (isRevertContractException(err)) 
+                return null;
+            
+            throw new NestedError(err, 'Unable to get the stage amount of the proposal.');
+        }
+    }
+
+    /**
+     * Returns status of the current challenge proposal
+     * @param {Address} address - The wallet address
+     * @param {Address} ct - The currency address
+     * @param {Integer} id - The currency id
+     * @returns {Promise} A promise that resolves into a string of either ['Qualified', 'Disqualified'] indicating the status of the latest challenge.
+     * @example
+     * let status = await driipSettlement.getCurrentProposalStatus(address, ct, id);
+     */
+    async getCurrentProposalStatus(address, ct, id) {
+        const challengeStatuses = ['Qualified', 'Disqualified'];
+        try {
+            const statusIndex = await this.proposalStatus(address, ct, id);
+            return challengeStatuses[statusIndex];
+        }
+        catch (err) {
+            if (isRevertContractException(err)) 
+                return null;
+            
+            throw new NestedError(err, 'Unable to get the status of the proposal.');
+        }
     }
 }
