+++ lib/settlement/index.js
@@ -1,6 +1,6 @@
 'use strict';
 
-const Settlement = require('./settlement');
+const Settlement = require('./settlement-factory');
 const DriipSettlement = require('./driip-settlement');
 const NullSettlement = require('./null-settlement');
