+++ lib/settlement/settlement-factory.spec.js
@@ -65,7 +65,7 @@ function proxyquireWallet() {
 }
 
 function proxyquireSettlement() {
-    return proxyquire('./settlement', {
+    return proxyquire('./settlement-factory', {
         './driip-settlement': function() {
             return stubbedDriipSettlement;
         },
@@ -77,7 +77,7 @@ function proxyquireSettlement() {
 
 function proxyquireSettlementModule() {
     return proxyquire('./index', {
-        './settlement': proxyquireSettlement()
+        './settlement-factory': proxyquireSettlement()
     });
 }
